refactor: move dialogs to features
