+++ frontend/src/features/Dialogs/index.ts
@@ -1,0 +1,1 @@
+export { useDialog, DialogProvider } from './useDialog';
