+++ frontend/src/hooks/index.ts
@@ -1,4 +1,2 @@
 export { default as useLocalStorage } from './useLocalStorage';
 export { default as useKeyDown } from './useKeyDown';
-
-export { DialogProvider, useDialog } from './useDialog';
